fix root/asset-dir functions and move them to utils
