+++ src/odm_validation/utils.py
@@ -1,8 +1,37 @@
-from os.path import splitext
+import inspect
+import os
+import sys
+from os.path import join, splitext
+from pathlib import Path
 
 import csv
 import json
 import yaml
+
+
+def get_pkg_dir() -> str:
+    '''returns the actual pkg. dir or `/src/odm_validation/` if in dev. env.'''
+    # this module file is located directly under the pkg. dir, so we can lookup
+    # its path and use that directly
+    mod = sys.modules[__name__]
+    mod_path = inspect.getfile(mod)
+    mod_dir = os.path.dirname(mod_path)
+    return mod_dir
+
+
+def get_asset_dir() -> str:
+    pkg_dir = get_pkg_dir()
+    asset_dir = join(pkg_dir, 'assets')
+
+    # If the `asset_dir` path doesn't exist, then we can assume that the
+    # package hasn't been installed, but is used directly in a development
+    # environment, meaning that we need to specify the path as it is in the
+    # repo.
+    if not os.path.isdir(asset_dir):
+        root_dir = Path(pkg_dir).parent.parent
+        asset_dir = join(root_dir, 'assets')
+
+    return asset_dir
 
 
 def import_csv_file(path: str) -> list[dict]:
